Refactor Store and Graph interfaces
This shifts around some of the interfaces between memory definitions and
the root interface definitions.
